+++ api_server/src/lib.rs
@@ -26,6 +26,7 @@ use std::sync::mpsc;
 use std::sync::{Arc, Mutex, RwLock};
 use std::{fmt, io};
 
+use futures::sync::oneshot;
 use futures::{Future, Stream};
 use hyper::server::Http;
 use tokio_core::reactor::Core;
@@ -36,8 +37,100 @@ use logger::{Metric, METRICS};
 use mmds::data_store::Mmds;
 use sys_util::EventFd;
 use vmm::default_syscalls;
+use vmm::vmm_config::boot_source::BootSourceConfig;
+use vmm::vmm_config::drive::BlockDeviceConfig;
 use vmm::vmm_config::instance_info::InstanceInfo;
-use vmm::VmmAction;
+use vmm::vmm_config::logger::LoggerConfig;
+use vmm::vmm_config::machine_config::VmConfig;
+use vmm::vmm_config::net::{NetworkInterfaceConfig, NetworkInterfaceUpdateConfig};
+use vmm::vmm_config::vsock::VsockDeviceConfig;
+use vmm::VmmActionError;
+
+/// This enum represents the public interface of the VMM. Each action contains various
+/// bits of information (ids, paths, etc.).
+#[derive(PartialEq)]
+pub enum VmmAction {
+    /// Configure the boot source of the microVM using as input the `ConfigureBootSource`. This
+    /// action can only be called before the microVM has booted.
+    ConfigureBootSource(BootSourceConfig),
+    /// Configure the logger using as input the `LoggerConfig`. This action can only be called
+    /// before the microVM has booted.
+    ConfigureLogger(LoggerConfig),
+    /// Get the configuration of the microVM.
+    GetVmConfiguration,
+    /// Flush the metrics. This action can only be called after the logger has been configured.
+    FlushMetrics,
+    /// Add a new block device or update one that already exists using the `BlockDeviceConfig` as
+    /// input. This action can only be called before the microVM has booted.
+    InsertBlockDevice(BlockDeviceConfig),
+    /// Add a new network interface config or update one that already exists using the
+    /// `NetworkInterfaceConfig` as input. This action can only be called before the microVM has
+    /// booted.
+    InsertNetworkDevice(NetworkInterfaceConfig),
+    /// Set the vsock device or update the one that already exists using the
+    /// `VsockDeviceConfig` as input. This action can only be called before the microVM has
+    /// booted.
+    SetVsockDevice(VsockDeviceConfig),
+    /// Update the size of an existing block device specified by an ID. The ID is the first data
+    /// associated with this enum variant. This action can only be called after the microVM is
+    /// started.
+    RescanBlockDevice(String),
+    /// Set the microVM configuration (memory & vcpu) using `VmConfig` as input. This
+    /// action can only be called before the microVM has booted.
+    SetVmConfiguration(VmConfig),
+    /// Launch the microVM. This action can only be called before the microVM has booted.
+    StartMicroVm,
+    /// Send CTRL+ALT+DEL to the microVM, using the i8042 keyboard function. If an AT-keyboard
+    /// driver is listening on the guest end, this can be used to shut down the microVM gracefully.
+    SendCtrlAltDel,
+    /// Update the path of an existing block device. The data associated with this variant
+    /// represents the `drive_id` and the `path_on_host`.
+    UpdateBlockDevicePath(String, String),
+    /// Update a network interface, after microVM start. Currently, the only updatable properties
+    /// are the RX and TX rate limiters.
+    UpdateNetworkInterface(NetworkInterfaceUpdateConfig),
+}
+
+/// The enum represents the response sent by the VMM in case of success. The response is either
+/// empty, when no data needs to be sent, or an internal VMM structure.
+#[derive(Debug)]
+pub enum VmmData {
+    /// No data is sent on the channel.
+    Empty,
+    /// The microVM configuration represented by `VmConfig`.
+    MachineConfiguration(VmConfig),
+}
+
+/// One shot channel used by VMM to send a response.
+pub type ResponseSender = oneshot::Sender<std::result::Result<VmmData, VmmActionError>>;
+/// One shot channel used to receive a response from the VMM.
+pub type ResponseReceiver = oneshot::Receiver<std::result::Result<VmmData, VmmActionError>>;
+
+/// Wrapper over requested action to be done by the VMM and sender where the response will go.
+/// This is wrapped in a struct to allow custom PartialEq.
+pub struct VmmRequest {
+    inner: Box<(VmmAction, ResponseSender)>,
+}
+
+impl VmmRequest {
+    /// Create a VmmRequest from given VmmAction and ResponseSender.
+    pub fn new(action: VmmAction, response_sender: ResponseSender) -> VmmRequest {
+        VmmRequest {
+            inner: Box::new((action, response_sender)),
+        }
+    }
+
+    pub fn unpack(self) -> (VmmAction, ResponseSender) {
+        let inner_box = *self.inner;
+        (inner_box.0, inner_box.1)
+    }
+}
+
+impl PartialEq for VmmRequest {
+    fn eq(&self, other: &VmmRequest) -> bool {
+        self.inner.0 == other.inner.0
+    }
+}
 
 pub enum Error {
     Io(io::Error),
@@ -70,7 +163,7 @@ pub struct ApiServer {
     // VMM instance info directly accessible from the API thread.
     vmm_shared_info: Arc<RwLock<InstanceInfo>>,
     // Sender which allows passing messages to the VMM.
-    api_request_sender: Rc<mpsc::Sender<Box<VmmAction>>>,
+    api_request_sender: Rc<mpsc::Sender<VmmRequest>>,
     efd: Rc<EventFd>,
 }
 
@@ -78,7 +171,7 @@ impl ApiServer {
     pub fn new(
         mmds_info: Arc<Mutex<Mmds>>,
         vmm_shared_info: Arc<RwLock<InstanceInfo>>,
-        api_request_sender: mpsc::Sender<Box<VmmAction>>,
+        api_request_sender: mpsc::Sender<VmmRequest>,
         kick_vmm_efd: EventFd,
     ) -> Result<Self> {
         Ok(ApiServer {