+++ api_server/src/request/boot_source.rs
@@ -6,9 +6,9 @@ use std::result;
 use futures::sync::oneshot;
 use hyper::Method;
 
+use super::{VmmAction, VmmRequest};
 use request::{IntoParsedRequest, ParsedRequest};
 use vmm::vmm_config::boot_source::BootSourceConfig;
-use vmm::VmmAction;
 
 impl IntoParsedRequest for BootSourceConfig {
     fn into_parsed_request(
@@ -18,7 +18,7 @@ impl IntoParsedRequest for BootSourceConfig {
     ) -> result::Result<ParsedRequest, String> {
         let (sender, receiver) = oneshot::channel();
         Ok(ParsedRequest::Sync(
-            Box::new(VmmAction::ConfigureBootSource(self, sender)),
+            VmmRequest::new(VmmAction::ConfigureBootSource(self), sender),
             receiver,
         ))
     }
@@ -42,7 +42,7 @@ mod tests {
         assert!(body
             .into_parsed_request(None, Method::Put)
             .eq(&Ok(ParsedRequest::Sync(
-                Box::new(VmmAction::ConfigureBootSource(same_body, sender)),
+                VmmRequest::new(VmmAction::ConfigureBootSource(same_body), sender),
                 receiver
             ))))
     }